+++ Mlib/Render/Text/Charsets.cpp
@@ -6,7 +6,7 @@ using namespace Mlib;
 std::u32string Mlib::ascii_chars() {
     std::u32string result(95, '?');
     for (auto&& [i, r] : enumerate(result)) {
-        r = i + 32;
+        r = (uint32_t)i + 32;
     }
     return result;
 }
